+++ nexus/nexus-proxy/src/main/java/org/sonatype/nexus/apachehttpclient/NonProxyHostsAwareHttpRoutePlanner.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.httpclient;
+package org.sonatype.nexus.apachehttpclient;
 
 import java.util.Set;
 import java.util.regex.Pattern;
