+++ nexus/nexus-proxy/src/test/java/org/sonatype/nexus/apachehttpclient/EvictingThreadTest.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.httpclient;
+package org.sonatype.nexus.apachehttpclient;
 
 import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.atLeastOnce;
